refactor(acp): migrate importers and delete bitrouter-substrate
The four consumers now import `bitrouter_sdk::acp::…` directly —
`acp_cli.rs` (`engine::Session`, `down::serve`, `telemetry`, `translate`),
`agents.rs` (`up::health_check`), `optimization/evaluator.rs`
(`engine`, `translate`) — so the transitional shims have no callers left.

`spawn.rs` was the last user of `dotdir::ensure_self_ignored`, which was
never ACP: it is a `bitrouter launch` concern, so the twenty lines and their
tests move into `spawn.rs` beside `launch_state_dir`, their only caller.

With nothing left in it, `crates/bitrouter-substrate` is deleted along with
its workspace member entry and the `apps/bitrouter` dependency.

Two identifiers still named the departed crate: the upstream connection's
thread name and the ACP `clientInfo` name it sends on `initialize`. Both
become `bitrouter-acp`. The `clientInfo` string is wire-visible to the agent
child (nothing branches on it) — flagged rather than hidden.

`docs/DEVELOPMENT.md` and the `/bitrouter` skill are updated in lockstep;
the dated `docs/*_SPEC.md` design records are left as written.
