+++ Cargo.toml
@@ -33,7 +33,6 @@ keyring = { version = "3", default-features = false, features = [
 bitrouter-sdk = { path = "crates/bitrouter-sdk", version = "1.0.0-alpha.27", default-features = false }
 bitrouter-cloud-sdk = { path = "crates/bitrouter-cloud-sdk", version = "1.0.0-alpha.27" }
 bitrouter-providers = { path = "crates/bitrouter-providers", version = "1.0.0-alpha.27" }
-bitrouter-substrate = { path = "crates/bitrouter-substrate", version = "1.0.0-alpha.27" }
 bitrouter-guardrails = { path = "crates/bitrouter-guardrails", version = "1.0.0-alpha.27" }
 bitrouter-observe = { path = "crates/bitrouter-observe", version = "1.0.0-alpha.27", default-features = false }
