+++ apps/bitrouter/Cargo.toml
@@ -18,7 +18,6 @@ path = "src/main.rs"
 
 [dependencies]
 bitrouter-sdk = { workspace = true, features = ["server", "config_file", "mcp", "acp"] }
-bitrouter-substrate.workspace = true
 bitrouter-cloud-sdk.workspace = true
 bitrouter-providers = { workspace = true, features = ["pkce"] }
 bitrouter-guardrails.workspace = true
