+++ apps/bitrouter/src/acp_cli.rs
@@ -43,9 +43,9 @@ use futures::StreamExt;
 use serde::Serialize;
 use tokio::io::{AsyncWrite, AsyncWriteExt};
 
-use bitrouter_substrate::engine::LaunchOptions;
-use bitrouter_substrate::telemetry::RequestCompleted;
-use bitrouter_substrate::translate::SessionUpdateKind;
+use bitrouter_sdk::acp::engine::LaunchOptions;
+use bitrouter_sdk::acp::telemetry::RequestCompleted;
+use bitrouter_sdk::acp::translate::SessionUpdateKind;
 
 use crate::paths::ConfigSource;
 
@@ -550,7 +550,7 @@ pub async fn serve(ctx: SpawnContext<'_>) -> Result<()> {
     let cwd = std::env::current_dir().context("resolving current directory")?;
     // Deferred open: the upstream `session/new` runs when the manager sends
     // its own `session/new`, so the manager's cwd + mcpServers are relayed.
-    let session = bitrouter_substrate::engine::Session::launch_deferred(
+    let session = bitrouter_sdk::acp::engine::Session::launch_deferred(
         &catalog,
         agent_id,
         cwd.clone(),
@@ -561,7 +561,7 @@ pub async fn serve(ctx: SpawnContext<'_>) -> Result<()> {
     let exporter = attach_observability(&config, agent_id, &session).await;
     let session = Arc::new(session);
 
-    let served = bitrouter_substrate::down::serve(Arc::clone(&session)).await;
+    let served = bitrouter_sdk::acp::down::serve(Arc::clone(&session)).await;
 
     // No manager left: shut the session down deliberately so the agent child
     // is reaped (same semantics as `prompt`). Once serving ends, the forwarding
@@ -626,7 +626,7 @@ where
 
     let catalog = catalog_from_config(&config)?;
     let cwd = std::env::current_dir().context("resolving current directory")?;
-    let session = bitrouter_substrate::engine::Session::launch(&catalog, agent_id, cwd, options)
+    let session = bitrouter_sdk::acp::engine::Session::launch(&catalog, agent_id, cwd, options)
         .await
         .with_context(|| format!("launching acp session for agent '{agent_id}'"))?;
     let exporter = attach_observability(&config, agent_id, &session).await;
@@ -693,7 +693,7 @@ where
 /// early-return in the `no_wait` branch above doesn't borrow `session` past its
 /// drop point.
 async fn prompt_wait<W>(
-    session: bitrouter_substrate::engine::Session,
+    session: bitrouter_sdk::acp::engine::Session,
     text: &str,
     contract: Option<crate::result_contract::ResultContract>,
     out: &mut W,
@@ -765,7 +765,7 @@ where
 /// Drive one prompt turn: stream its updates to `out` (accumulating message
 /// text when `capture`), and return the typed response plus the reply text.
 async fn run_turn<W>(
-    session: &bitrouter_substrate::engine::Session,
+    session: &bitrouter_sdk::acp::engine::Session,
     updates: &mut (impl futures::Stream<Item = SessionUpdateKind> + Unpin),
     text: &str,
     capture: bool,
@@ -841,7 +841,7 @@ where
 async fn attach_observability(
     config: &Config,
     agent_id: &str,
-    session: &bitrouter_substrate::engine::Session,
+    session: &bitrouter_sdk::acp::engine::Session,
 ) -> Option<Arc<bitrouter_observe::otel::OtelExporter>> {
     let exporter = crate::assemble::build_otel_exporter_standalone(config).await;
     let recorder = exporter.as_ref().map(|exporter| {
@@ -875,7 +875,7 @@ async fn attach_observability(
     if let Some(recorder) = recorder {
         let mut updates = session.updates();
         tokio::spawn(async move {
-            use bitrouter_substrate::translate::ToolStatus;
+            use bitrouter_sdk::acp::translate::ToolStatus;
             while let Some(update) = updates.next().await {
                 match update {
                     SessionUpdateKind::ToolCall {
