+++ apps/bitrouter/src/agents.rs
@@ -109,7 +109,7 @@ pub async fn check(config: &Config) -> Vec<CheckRow> {
     for (id, cfg) in sorted {
         let outcome = match &cfg.transport {
             AcpTransport::Stdio { command, args, env } => {
-                bitrouter_substrate::up::health_check(command, args, env).await
+                bitrouter_sdk::acp::up::health_check(command, args, env).await
             }
         };
         out.push(CheckRow {
