--- crates/bitrouter-substrate/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "bitrouter-substrate"
-description = "Transitional shim: the ACP thin proxy now lives in bitrouter-sdk::acp"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-bitrouter-sdk = { workspace = true, features = ["acp"] }
-
-[dev-dependencies]
-tempfile.workspace = true
